Adjustments to author, logo, menu, col. Introduced button, checkbox, footer, feed, testing page
I made adjustments to existing components (Author, Logo, Menu, Col) and started working on the Button and Checkbox components. State transitions are working on buttons as well as using them to go to links. For MVP I need to implement an overlay functionality and the ability to use them as a form submit input.